+++ example/lib/customs/pickers/directory_file_asset_picker.dart
@@ -381,8 +381,8 @@ class FileAssetPickerBuilder
   @override
   Future<void> viewAsset(
     BuildContext context,
-    int index,
-    AssetEntity currentAsset,
+    int? index,
+    File currentAsset,
   ) async {
     final List<File>? result = await Navigator.of(context).push<List<File>?>(
       PageRouteBuilder<List<File>>(
@@ -393,7 +393,8 @@ class FileAssetPickerBuilder
         ) {
           return AssetPickerViewer<File, Directory>(
             builder: FileAssetPickerViewerBuilderDelegate(
-              currentIndex: index,
+              currentIndex:
+                  index ?? provider.selectedAssets.indexOf(currentAsset),
               previewAssets: provider.selectedAssets,
               provider: FileAssetPickerViewerProvider(provider.selectedAssets),
               themeData: AssetPicker.themeData(themeColor),
@@ -506,15 +507,11 @@ class FileAssetPickerBuilder
   PreferredSizeWidget appBar(BuildContext context) {
     final AppBar appBar = AppBar(
       backgroundColor: theme.appBarTheme.backgroundColor,
-      centerTitle: isAppleOS(context),
-      title: pathEntitySelector(context),
+      title: Semantics(
+        onTapHint: semanticsTextDelegate.sActionSwitchPathLabel,
+        child: pathEntitySelector(context),
+      ),
       leading: backButton(context),
-      actions: !isAppleOS(context)
-          ? <Widget>[
-              confirmButton(context),
-              const SizedBox(width: 14.0),
-            ]
-          : null,
     );
     appBarPreferredSize ??= appBar.preferredSize;
     return appBar;
@@ -1386,24 +1383,51 @@ class FileAssetPickerViewerBuilderDelegate
         color: themeData.canvasColor.withOpacity(0.85),
         child: Row(
           children: <Widget>[
-            const BackButton(),
-            if (!isAppleOS(context))
-              StreamBuilder<int>(
-                initialData: currentIndex,
-                stream: pageStreamController.stream,
-                builder: (BuildContext _, AsyncSnapshot<int> snapshot) {
-                  return Text(
-                    '${snapshot.data! + 1}/${previewAssets.length}',
-                    style: const TextStyle(
-                      fontSize: 18.0,
-                      fontWeight: FontWeight.bold,
-                    ),
-                  );
-                },
+            Expanded(
+              child: Align(
+                alignment: AlignmentDirectional.centerStart,
+                child: Semantics(
+                  sortKey: ordinalSortKey(0),
+                  child: IconButton(
+                    icon: const Icon(Icons.close),
+                    tooltip: MaterialLocalizations.of(
+                      context,
+                    ).backButtonTooltip,
+                    onPressed: Navigator.of(context).maybePop,
+                  ),
+                ),
               ),
-            const Spacer(),
-            if (isAppleOS(context) && provider != null) selectButton(context),
-            if (!isAppleOS(context) && provider != null) confirmButton(context),
+            ),
+            Expanded(
+              child: Center(
+                child: StreamBuilder<int>(
+                  initialData: currentIndex,
+                  stream: pageStreamController.stream,
+                  builder: (BuildContext _, AsyncSnapshot<int> snapshot) {
+                    return Text(
+                      '${snapshot.requireData + 1}/${previewAssets.length}',
+                      style: const TextStyle(
+                        fontSize: 18.0,
+                        fontWeight: FontWeight.bold,
+                      ),
+                    );
+                  },
+                ),
+              ),
+            ),
+            if (provider != null)
+              Expanded(
+                child: Container(
+                  alignment: AlignmentDirectional.centerEnd,
+                  padding: const EdgeInsetsDirectional.only(end: 14),
+                  child: Semantics(
+                    sortKey: ordinalSortKey(0.2),
+                    child: selectButton(context),
+                  ),
+                ),
+              )
+            else
+              const Spacer(),
           ],
         ),
       ),