+++ lib/src/constants/typedefs.dart
@@ -6,6 +6,10 @@ import 'dart:async';
 
 import 'package:flutter/widgets.dart';
 import 'package:photo_manager/photo_manager.dart' show PermissionState;
+import 'package:provider/provider.dart';
+
+/// Mirroring [ChangeNotifierProvider].
+typedef CNP<T extends ChangeNotifier?> = ChangeNotifierProvider<T>;
 
 /// {@template wechat_assets_picker.LoadingIndicatorBuilder}
 /// Build the loading indicator with the given [isAssetsEmpty].