+++ lib/src/widget/asset_picker_app_bar.dart
@@ -6,9 +6,11 @@ import 'dart:ui' as ui;
 
 import 'package:flutter/material.dart';
 import 'package:flutter/semantics.dart';
-
+ 
 import '../constants/extensions.dart';
-
+import 'package:flutter/services.dart';
+import 'package:wechat_picker_library/wechat_picker_library.dart';
+ 
 /// A custom app bar.
 /// 自定义的顶栏
 class AssetPickerAppBar extends StatelessWidget implements PreferredSizeWidget {
@@ -101,8 +103,10 @@ class AssetPickerAppBar extends StatelessWidget implements PreferredSizeWidget {
   @override
   Widget build(BuildContext context) {
     final ThemeData theme = Theme.of(context);
-    final IconThemeData iconTheme = this.iconTheme ?? theme.iconTheme;
-    final Widget? titleWidget;
+     final IconThemeData iconTheme = this.iconTheme ?? theme.iconTheme;
+     final AppBarTheme appBarTheme = theme.appBarTheme;
+
+     final Widget? titleWidget;
     if (centerTitle) {
       titleWidget = Center(child: title);
     } else {
@@ -118,7 +122,10 @@ class AssetPickerAppBar extends StatelessWidget implements PreferredSizeWidget {
             PositionedDirectional(
               top: 0.0,
               bottom: 0.0,
-              child: leading ?? const BackButton(),
+              child: IconTheme.merge(
+                data: appBarTheme.iconTheme ?? theme.iconTheme,
+                child: leading ?? const BackButton(),
+              ),
             ),
           if (titleWidget != null)
             PositionedDirectional(
@@ -131,7 +138,8 @@ class AssetPickerAppBar extends StatelessWidget implements PreferredSizeWidget {
                     ? Alignment.center
                     : AlignmentDirectional.centerStart,
                 child: DefaultTextStyle(
-                  style: theme.textTheme.titleLarge!.copyWith(fontSize: 23.0),
+                  style: appBarTheme.titleTextStyle ??
+                      theme.textTheme.titleLarge!.copyWith(fontSize: 23.0),
                   maxLines: 1,
                   softWrap: false,
                   overflow: TextOverflow.ellipsis,
@@ -146,9 +154,15 @@ class AssetPickerAppBar extends StatelessWidget implements PreferredSizeWidget {
               top: 0.0,
               end: 0.0,
               height: _barHeight,
-              child: Padding(
-                padding: actionsPadding ?? EdgeInsets.zero,
-                child: Row(mainAxisSize: MainAxisSize.min, children: actions!),
+              child: IconTheme.merge(
+                data: appBarTheme.actionsIconTheme ?? theme.iconTheme,
+                child: Padding(
+                  padding: actionsPadding ?? EdgeInsets.zero,
+                  child: Row(
+                    mainAxisSize: MainAxisSize.min,
+                    children: actions!,
+                  ),
+                ),
               ),
             ),
         ],
@@ -173,18 +187,21 @@ class AssetPickerAppBar extends StatelessWidget implements PreferredSizeWidget {
     }
 
     /// Apply the icon theme data.
-    child = IconTheme.merge(data: iconTheme, child: child);
+    child = IconTheme.merge(
+      data: iconTheme ?? appBarTheme.iconTheme ?? theme.iconTheme,
+      child: child,
+    );
 
-    final Color effectiveBackgroundColor =
-        backgroundColor ?? theme.colorScheme.surface;
-
-    child = Column(
+     child = Column(
       mainAxisSize: MainAxisSize.min,
       children: <Widget>[
         child,
         if (bottom != null) bottom!,
-      ],
+      ], 
 //     // Set [SystemUiOverlayStyle] according to the brightness.
+//     final Color effectiveBackgroundColor = backgroundColor ??
+//         appBarTheme.backgroundColor ??
+//         theme.colorScheme.surface;
 //     final Brightness effectiveBrightness = brightness ??
 //         appBarTheme.systemOverlayStyle?.statusBarBrightness ??
 //         theme.brightness;
@@ -198,10 +215,10 @@ class AssetPickerAppBar extends StatelessWidget implements PreferredSizeWidget {
 //     child = AnnotatedRegion<SystemUiOverlayStyle>(
 //       value: overlayStyle,
 //       child: child,
-    );
+     );
 
     final Widget result = Material(
-      // Wrap to ensure the child rendered correctly
+      // Wrap to ensure the child rendered correctly.
       color: Color.lerp(
         effectiveBackgroundColor,
         Colors.transparent,
@@ -211,7 +228,11 @@ class AssetPickerAppBar extends StatelessWidget implements PreferredSizeWidget {
       child: child,
     );
     return semanticsBuilder?.call(result) ??
-        Semantics(sortKey: const OrdinalSortKey(0), child: result);
+        Semantics(
+          sortKey: const OrdinalSortKey(0),
+          explicitChildNodes: true,
+          child: result,
+        );
   }
 }