+++ lib/src/widget/asset_picker_viewer.dart
@@ -40,8 +40,11 @@ class AssetPickerViewer<Asset, Path> extends StatefulWidget {
     int? maxAssets,
     bool shouldReversePreview = false,
     AssetSelectPredicate<AssetEntity>? selectPredicate,
+    PermissionRequestOption permissionRequestOption =
+        const PermissionRequestOption(),
   }) async {
-    // await AssetPicker.permissionCheck();
+
+//     await AssetPicker.permissionCheck(requestOption: permissionRequestOption);
     final Widget viewer = AssetPickerViewer<AssetEntity, AssetPathEntity>(
       builder: DefaultAssetPickerViewerBuilderDelegate(
         currentIndex: currentIndex,
@@ -81,8 +84,12 @@ class AssetPickerViewer<Asset, Path> extends StatefulWidget {
   static Future<List<A>?> pushToViewerWithDelegate<A, P>(
     BuildContext context, {
     required AssetPickerViewerBuilderDelegate<A, P> delegate,
+    PermissionRequestOption permissionRequestOption =
+        const PermissionRequestOption(),
   }) async {
-    // await AssetPicker.permissionCheck();
+
+//     await AssetPicker.permissionCheck(requestOption: permissionRequestOption);
+
     final Widget viewer = AssetPickerViewer<A, P>(builder: delegate);
     final PageRouteBuilder<List<A>> pageRoute = PageRouteBuilder<List<A>>(
       pageBuilder: (_, __, ___) => viewer,